fix fetching data and few icons
