+++ components/Icon.tsx
@@ -1,0 +1,40 @@
+import React from "react";
+import iconMap from "./icons/icon-map";
+
+const EmptyIcon = () => <div />;
+
+/**
+ * Generic icon
+ * @param name - icon name as defined in icon-map.js
+ * @param size - icon size (width and height)
+ * @param color - icon color (hexadecimal or TailwindCSS color)
+ * @param className - additional TailwindCSS classes
+ * @param rest - additional props
+ * @returns {JSX.Element} - icon component
+ */
+
+interface IconProps {
+    name?: string,
+    size?: string,
+    color?: string,
+    className?: string 
+}
+const Icon:React.FC<IconProps> = ({ name, size, color, className, ...rest }) => {
+// @ts-ignore
+  const Icon = iconMap[name] || EmptyIcon;
+  return (
+    <Icon
+      color={color}
+      className={className}
+      style={{ width: size, height: size }}
+      {...rest}
+    />
+  );
+};
+
+Icon.defaultProps = {
+  size: "32px",
+  color: "black",
+};
+
+export default Icon;
