--- components/icons/AppsIcon.tsx
@@ -1,56 +1,0 @@
-import React from "react";
-
-interface AppsIconProps {
-  color: string;
-}
-
-const AppsIcon: React.FC<AppsIconProps> = ({ color }) => {
-  return (
-    <svg
-      width="35"
-      height="35"
-      viewBox="0 0 35 35"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M20.9 12.5H14.1C13.2163 12.5 12.5 13.2163 12.5 14.1V20.9C12.5 21.7837 13.2163 22.5 14.1 22.5H20.9C21.7837 22.5 22.5 21.7837 22.5 20.9V14.1C22.5 13.2163 21.7837 12.5 20.9 12.5Z"
-        fill={color}
-      />
-      <path
-        d="M20.9 25H14.1C13.2163 25 12.5 25.7163 12.5 26.6V33.4C12.5 34.2837 13.2163 35 14.1 35H20.9C21.7837 35 22.5 34.2837 22.5 33.4V26.6C22.5 25.7163 21.7837 25 20.9 25Z"
-        fill={color}
-      />
-      <path
-        d="M27.925 0H26.6C26.1757 0 25.7687 0.168571 25.4686 0.468629C25.1686 0.768687 25 1.17565 25 1.6V8.4C25 8.82435 25.1686 9.23131 25.4686 9.53137C25.7687 9.83143 26.1757 10 26.6 10H33.4C33.8243 10 34.2313 9.83143 34.5314 9.53137C34.8314 9.23131 35 8.82435 35 8.4V7.075C35 5.19859 34.2546 3.39904 32.9278 2.07222C31.601 0.7454 29.8014 0 27.925 0Z"
-        fill={color}
-      />
-      <path
-        d="M8.4 12.5H1.6C0.716344 12.5 0 13.2163 0 14.1V20.9C0 21.7837 0.716344 22.5 1.6 22.5H8.4C9.28366 22.5 10 21.7837 10 20.9V14.1C10 13.2163 9.28366 12.5 8.4 12.5Z"
-        fill={color}
-      />
-      <path
-        d="M7.075 35H8.40625C8.82951 34.9983 9.23487 34.829 9.53358 34.5292C9.83229 34.2293 10 33.8233 10 33.4V26.6C10 26.1757 9.83143 25.7687 9.53137 25.4686C9.23131 25.1686 8.82435 25 8.4 25H1.6C1.17565 25 0.768687 25.1686 0.468629 25.4686C0.168571 25.7687 0 26.1757 0 26.6V27.9312C0.00165665 29.8066 0.747787 31.6045 2.07443 32.93C3.40107 34.2555 5.19968 35 7.075 35Z"
-        fill={color}
-      />
-      <path
-        d="M1.6 10H8.4C8.82435 10 9.23131 9.83143 9.53137 9.53137C9.83143 9.23131 10 8.82435 10 8.4V1.6C10 1.17565 9.83143 0.768687 9.53137 0.468629C9.23131 0.168571 8.82435 0 8.4 0H7.075C5.19859 0 3.39904 0.7454 2.07222 2.07222C0.7454 3.39904 0 5.19859 0 7.075V8.40625C0.00165338 8.82951 0.170954 9.23487 0.470833 9.53358C0.770712 9.83229 1.17673 10 1.6 10Z"
-        fill={color}
-      />
-      <path
-        d="M33.4 25H26.6C26.1757 25 25.7687 25.1686 25.4686 25.4686C25.1686 25.7687 25 26.1757 25 26.6V33.4C25 33.8243 25.1686 34.2313 25.4686 34.5314C25.7687 34.8314 26.1757 35 26.6 35H27.9312C29.8066 34.9983 31.6045 34.2522 32.93 32.9256C34.2555 31.5989 35 29.8003 35 27.925V26.6C35 26.1757 34.8314 25.7687 34.5314 25.4686C34.2313 25.1686 33.8243 25 33.4 25Z"
-        fill={color}
-      />
-      <path
-        d="M20.9 0H14.1C13.2163 0 12.5 0.716344 12.5 1.6V8.4C12.5 9.28366 13.2163 10 14.1 10H20.9C21.7837 10 22.5 9.28366 22.5 8.4V1.6C22.5 0.716344 21.7837 0 20.9 0Z"
-        fill={color}
-      />
-      <path
-        d="M33.4 12.5H26.6C25.7163 12.5 25 13.2163 25 14.1V20.9C25 21.7837 25.7163 22.5 26.6 22.5H33.4C34.2837 22.5 35 21.7837 35 20.9V14.1C35 13.2163 34.2837 12.5 33.4 12.5Z"
-        fill={color}
-      />
-    </svg>
-  );
-};
-
-export default AppsIcon;
