--- components/icons/DollarIcon.tsx
@@ -1,23 +1,0 @@
-import React from "react";
-
-const DollarIcon = () => {
-  return (
-    <svg
-      width="24"
-      height="25"
-      viewBox="0 0 24 25"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <rect y="0.5" width="24" height="24" rx="12" fill="white" />
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M24 12.5C24 19.15 18.65 24.5 12 24.5C5.34996 24.5 0 19.15 0 12.5C0 5.84996 5.34996 0.5 12 0.5C18.65 0.5 24 5.84996 24 12.5ZM5.00004 10.0499C3.54996 13.9 5.55 18.25 9.45 19.6499C9.6 19.75 9.75 19.9499 9.75 20.0999V20.8C9.75 20.9 9.75 20.95 9.69996 20.9999C9.65004 21.2 9.45 21.2999 9.24996 21.2C6.45 20.3 4.29996 18.1499 3.39996 15.35C1.89996 10.6 4.5 5.54992 9.24996 4.04992C9.3 4 9.39996 4 9.45 4C9.65004 4.04992 9.75 4.19992 9.75 4.39996V5.09992C9.75 5.35 9.65004 5.5 9.45 5.59996C7.40004 6.34996 5.75004 7.94992 5.00004 10.0499ZM14.3 4.24996C14.35 4.04992 14.55 3.94996 14.75 4.04992C17.5 4.94992 19.7 7.09996 20.6 9.94996C22.1 14.6999 19.5 19.75 14.75 21.25C14.7 21.2999 14.6 21.2999 14.55 21.2999C14.35 21.25 14.25 21.1 14.25 20.9V20.2C14.25 19.9499 14.35 19.7999 14.55 19.7C16.6 18.95 18.25 17.35 19 15.25C20.45 11.3999 18.45 7.04992 14.55 5.65C14.4 5.54992 14.25 5.35 14.25 5.14996V4.45C14.25 4.34992 14.25 4.3 14.3 4.24996ZM12.15 11.8C14.25 12.05 15.3 12.65 15.3 14.3999C15.3 15.7499 14.3 16.7999 12.8 17.05V18.25C12.75 18.5 12.6 18.65 12.4 18.65H11.65C11.4 18.5999 11.25 18.4499 11.25 18.25V17.05C9.6 16.7999 8.79996 15.8999 8.60004 14.65V14.6C8.60004 14.3999 8.75004 14.2499 8.94996 14.2499H9.80004C9.95004 14.2499 10.1 14.35 10.15 14.5499C10.3 15.2999 10.75 15.85 12.05 15.85C13 15.85 13.7 15.3 13.7 14.5C13.7 13.7 13.25 13.4 11.85 13.15C9.75 12.8999 8.75004 12.25 8.75004 10.6C8.75004 9.34996 9.69996 8.35 11.2 8.14996V7C11.25 6.74992 11.4 6.59992 11.6 6.59992H12.35C12.6 6.64996 12.75 6.79996 12.75 7V8.2C13.9 8.29996 14.8 9.14992 15 10.25V10.3C15 10.4999 14.85 10.6499 14.65 10.6499H13.85C13.7 10.6499 13.55 10.55 13.5 10.4C13.25 9.64996 12.75 9.34996 11.85 9.34996C10.85 9.34996 10.35 9.79984 10.35 10.4999C10.35 11.2 10.65 11.6 12.15 11.8Z"
-        fill="#2775CA"
-      />
-    </svg>
-  );
-};
-
-export default DollarIcon;
