--- components/icons/ExportIcon.tsx
@@ -1,24 +1,0 @@
-import React from "react";
-
-const ExportIcon = () => {
-  return (
-    <svg
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M23.2494 18.3756L23.2494 14.7643C23.2494 14.4292 23.1162 14.1077 22.8792 13.8707C22.6422 13.6337 22.3208 13.5006 21.9856 13.5006L20.7594 13.5006C20.4249 13.5016 20.1044 13.6352 19.8682 13.8721C19.632 14.1089 19.4994 14.4298 19.4994 14.7643L19.4994 16.7406C19.4984 17.0751 19.3648 17.3956 19.1279 17.6318C18.891 17.868 18.5702 18.0006 18.2356 18.0006L7.26314 18.0006C6.92797 18.0006 6.60653 17.8674 6.36953 17.6304C6.13253 17.3934 5.99939 17.072 5.99939 16.7368L5.99939 14.7643C5.99939 14.4292 5.86625 14.1077 5.62925 13.8707C5.39225 13.6337 5.07081 13.5006 4.73564 13.5006L3.50939 13.5006C3.17487 13.5016 2.85439 13.6352 2.61821 13.8721C2.38202 14.1089 2.24939 14.4298 2.24939 14.7643L2.24939 18.3756C2.24939 19.2707 2.60497 20.1291 3.2379 20.7621C3.87084 21.395 4.72928 21.7506 5.62439 21.7506L19.8744 21.7506C20.7695 21.7506 21.6279 21.395 22.2609 20.7621C22.8938 20.1291 23.2494 19.2707 23.2494 18.3756Z"
-        fill="#80858E"
-      />
-      <path
-        d="M7.38693 6.99059C7.34826 7.02924 7.31759 7.07514 7.29667 7.12565C7.27574 7.17615 7.26497 7.23029 7.26497 7.28496C7.26497 7.33964 7.27574 7.39377 7.29667 7.44428C7.31759 7.49479 7.34826 7.54069 7.38693 7.57934L8.92443 9.11309C8.96308 9.15175 9.00897 9.18243 9.05948 9.20335C9.10999 9.22428 9.16413 9.23505 9.2188 9.23505C9.27348 9.23505 9.32761 9.22428 9.37812 9.20335C9.42863 9.18243 9.47452 9.15175 9.51318 9.11309L10.9794 7.64684C11.002 7.62489 11.0304 7.61 11.0613 7.60402C11.0922 7.59804 11.1242 7.60123 11.1532 7.61319C11.1823 7.62515 11.2073 7.64535 11.2251 7.67132C11.2428 7.69729 11.2526 7.72789 11.2532 7.75934L11.2532 15.3343C11.2532 15.4441 11.2965 15.5494 11.3738 15.6274C11.451 15.7053 11.5559 15.7496 11.6657 15.7506L13.8332 15.7506C13.9436 15.7506 14.0494 15.7067 14.1275 15.6287C14.2056 15.5506 14.2494 15.4447 14.2494 15.3343L14.2494 7.75934C14.25 7.72789 14.2598 7.69729 14.2775 7.67132C14.2953 7.64535 14.3203 7.62515 14.3494 7.61319C14.3785 7.60123 14.4104 7.59804 14.4413 7.60402C14.4722 7.61 14.5006 7.62489 14.5232 7.64684L15.9894 9.11309C16.0281 9.15175 16.074 9.18243 16.1245 9.20335C16.175 9.22428 16.2291 9.23505 16.2838 9.23505C16.3385 9.23505 16.3926 9.22428 16.4431 9.20335C16.4936 9.18243 16.5395 9.15175 16.5782 9.11309L18.1119 7.57559C18.1506 7.53694 18.1813 7.49104 18.2022 7.44053C18.2231 7.39002 18.2339 7.33589 18.2339 7.28122C18.2339 7.22654 18.2231 7.17241 18.2022 7.1219C18.1813 7.07139 18.1506 7.02549 18.1119 6.98684L13.8107 2.68934C13.6714 2.54987 13.5059 2.43924 13.3238 2.36375C13.1417 2.28826 12.9466 2.24941 12.7494 2.24941C12.5523 2.24941 12.3571 2.28826 12.175 2.36375C11.9929 2.43924 11.8275 2.54987 11.6882 2.68934L7.38693 6.99059Z"
-        fill="#80858E"
-      />
-    </svg>
-  );
-};
-
-export default ExportIcon;
