--- components/icons/ImportIcon.tsx
@@ -1,24 +1,0 @@
-import React from "react";
-
-const ImportIcon = () => {
-  return (
-    <svg
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M6.43875 22.5L17.5612 22.5C18.8708 22.499 20.1264 21.9784 21.0524 21.0524C21.9784 20.1264 22.499 18.8708 22.5 17.5612L22.5 6.43875C22.499 5.12922 21.9784 3.8736 21.0524 2.94762C20.1264 2.02164 18.8708 1.50099 17.5613 1.5C17.1634 1.5 16.7819 1.65803 16.5006 1.93934C16.2193 2.22064 16.0613 2.60217 16.0613 3C16.0613 3.39782 16.2193 3.77936 16.5006 4.06066C16.7819 4.34196 17.1634 4.5 17.5613 4.5C17.8159 4.5 18.068 4.55015 18.3032 4.64758C18.5384 4.74501 18.7521 4.88782 18.9322 5.06785C19.2957 5.43143 19.5 5.92456 19.5 6.43875L19.5 16.5L4.5 16.5L4.5 6.43875C4.5 5.92456 4.70426 5.43143 5.06785 5.06785C5.43143 4.70426 5.92456 4.5 6.43875 4.5C6.83658 4.5 7.21811 4.34196 7.49941 4.06066C7.78072 3.77935 7.93875 3.39782 7.93875 3C7.93875 2.60217 7.78072 2.22064 7.49941 1.93934C7.21811 1.65803 6.83658 1.5 6.43875 1.5C5.12922 1.50099 3.87361 2.02164 2.94763 2.94762C2.02164 3.8736 1.501 5.12921 1.5 6.43875L1.5 17.5612C1.50099 18.8708 2.02164 20.1264 2.94762 21.0524C3.8736 21.9784 5.12922 22.499 6.43875 22.5Z"
-        fill="#80858E"
-      />
-      <path
-        d="M12 1.5C11.6022 1.5 11.2207 1.65804 10.9394 1.93934C10.6581 2.22064 10.5 2.60217 10.5 3L10.5 8.3775L9.56627 7.44375C9.4269 7.30438 9.26145 7.19383 9.07936 7.11841C8.89727 7.04298 8.70211 7.00417 8.50502 7.00417C8.30792 7.00417 8.11276 7.04298 7.93067 7.11841C7.74858 7.19383 7.58313 7.30438 7.44377 7.44375C7.3044 7.58311 7.19385 7.74856 7.11843 7.93065C7.043 8.11274 7.00418 8.30791 7.00418 8.505C7.00418 8.70209 7.043 8.89725 7.11843 9.07934C7.19385 9.26143 7.3044 9.42688 7.44377 9.56625L10.9388 13.0612C11.0781 13.2007 11.2435 13.3114 11.4256 13.3868C11.6077 13.4623 11.8029 13.5012 12 13.5012C12.1971 13.5012 12.3923 13.4623 12.5744 13.3868C12.7565 13.3114 12.922 13.2007 13.0613 13.0612L16.5563 9.56625C16.6956 9.42688 16.8062 9.26143 16.8816 9.07934C16.957 8.89725 16.9959 8.70209 16.9959 8.505C16.9959 8.30791 16.957 8.11275 16.8816 7.93066C16.8062 7.74857 16.6956 7.58311 16.5563 7.44375C16.4169 7.30438 16.2515 7.19383 16.0694 7.11841C15.8873 7.04299 15.6921 7.00417 15.495 7.00417C15.2979 7.00417 15.1028 7.04299 14.9207 7.11841C14.7386 7.19383 14.5731 7.30438 14.4338 7.44375L13.5 8.3775L13.5 3C13.5 2.60217 13.342 2.22064 13.0607 1.93934C12.7794 1.65804 12.3978 1.5 12 1.5Z"
-        fill="#80858E"
-      />
-    </svg>
-  );
-};
-
-export default ImportIcon;
