--- components/icons/MaticIcon.tsx
@@ -1,23 +1,0 @@
-import React from "react";
-
-const MaticIcon = () => {
-  return (
-    <svg
-      width="24"
-      height="25"
-      viewBox="0 0 24 25"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <rect y="0.5" width="24" height="24" rx="12" fill="white" />
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M24 12.5C24 19.1274 18.6274 24.5 12 24.5C5.37258 24.5 0 19.1274 0 12.5C0 5.87258 5.37258 0.5 12 0.5C18.6274 0.5 24 5.87258 24 12.5ZM16.0345 19.358L15.9325 19.3994L15.8186 19.5H15.5246C15.1969 19.5 15.1532 19.4767 15.0762 19.4358L15.0746 19.4349C14.9966 19.3935 11.679 17.4527 11.625 17.4113C11.5335 17.3446 11.4526 17.2648 11.385 17.1746C11.3198 17.0864 11.2692 16.9885 11.2351 16.8846C11.2284 16.862 11.2228 16.8431 11.218 16.824C11.1931 16.7233 11.1931 16.6164 11.1931 15.9201V15.1568L11.979 14.7012L12.7649 14.2456V15.003C12.7706 15.791 12.7709 15.7919 12.8122 15.9175C12.8143 15.9239 12.8166 15.9307 12.8189 15.9379C12.8893 16.14 13.0186 16.317 13.1909 16.4468L14.1567 17.0385C14.2874 17.1192 14.4062 17.1931 14.5125 17.2592C14.8892 17.4935 15.1092 17.6302 15.1466 17.6302C15.2834 17.6904 15.4317 17.7215 15.5815 17.7215C15.7314 17.7215 15.8796 17.6904 16.0165 17.6302C16.1545 17.571 17.8942 16.5414 17.9542 16.5C18.1695 16.3407 18.3136 16.1052 18.3562 15.8432V13.4231C18.3142 13.1684 18.1772 12.9384 17.9722 12.7781C17.9182 12.7189 16.1125 11.6302 16.0405 11.5947C16.0028 11.5737 15.9625 11.5578 15.9205 11.5474C15.661 11.4525 15.3741 11.4631 15.1226 11.5769L13.9227 12.3047C12.8129 12.9734 12.6329 13.074 11.979 13.4349C11.8897 13.4842 11.807 13.5297 11.7281 13.5731C11.229 13.8475 10.8856 14.0364 10.0412 14.5473L8.8414 15.2515C8.69779 15.318 8.54109 15.3524 8.38245 15.3524C8.2238 15.3524 8.06712 15.318 7.92351 15.2515C7.83952 15.2101 4.50994 13.2751 4.43194 13.2278C4.32325 13.1462 4.22811 13.0484 4.14998 12.9379C4.07419 12.8274 4.02307 12.7024 4 12.571V10.3994V8.22781C4.02294 8.09964 4.07131 7.97722 4.14234 7.86755C4.21338 7.75788 4.30569 7.66315 4.41395 7.58876C4.44394 7.56509 5.23585 7.10356 6.17773 6.55918C7.44957 5.82545 7.91151 5.56508 7.9775 5.5355L8.06749 5.5H8.66742H8.72141L8.84738 5.55917C9.00336 5.64201 12.267 7.53551 12.327 7.57693C12.497 7.70625 12.626 7.88072 12.6989 8.07987C12.7096 8.1126 12.7181 8.13772 12.725 8.16328C12.7529 8.26713 12.7529 8.37847 12.7529 9.03847V9.77811L11.967 10.2337L11.1691 10.6893V9.94379V9.20415V9.11538C11.1457 9.02403 11.1093 8.93639 11.0611 8.85502C10.9871 8.72227 10.8822 8.60872 10.7551 8.52367C10.6232 8.44674 8.88339 7.43492 8.8234 7.40533C8.68417 7.34213 8.53271 7.30941 8.37944 7.30941C8.22618 7.30941 8.07474 7.34213 7.9355 7.40533C7.90401 7.42949 7.30315 7.78375 6.79351 8.08422C6.42934 8.29893 6.11174 8.48618 6.08174 8.50591C5.98698 8.57078 5.90386 8.65079 5.83577 8.74261C5.7677 8.83947 5.71322 8.94496 5.67379 9.05622C5.59766 9.51189 5.58354 9.97555 5.63179 10.4349C5.61036 10.8291 5.61036 11.2241 5.63179 11.6183C5.66213 11.7532 5.717 11.8815 5.79377 11.9971C5.86292 12.0914 5.94581 12.1751 6.03974 12.2456C6.13573 12.3166 7.92951 13.3758 7.9955 13.3758C8.13377 13.4347 8.28284 13.4648 8.43346 13.4645C8.58553 13.4663 8.73578 13.4319 8.87138 13.3639C8.91338 13.3521 9.47131 13.0444 10.0712 12.6775L12.015 11.5533L13.9827 10.4112C14.2533 10.2551 14.4743 10.1253 14.6507 10.0217C15.0023 9.81534 15.1766 9.71302 15.2126 9.71302C15.3474 9.6527 15.4936 9.6215 15.6415 9.6215C15.7895 9.6215 15.9358 9.6527 16.0705 9.71302L19.2801 11.5651C19.4443 11.6481 19.5958 11.7537 19.73 11.8787C19.8734 12.0294 19.9674 12.219 20 12.4231V16.6539C19.9727 16.8379 19.889 17.0092 19.76 17.145C19.6873 17.2231 19.6022 17.289 19.5081 17.3403L17.7442 18.3639L16.0345 19.358Z"
-        fill="#854CE6"
-      />
-    </svg>
-  );
-};
-
-export default MaticIcon;
