--- components/icons/RampIcon.tsx
@@ -1,30 +1,0 @@
-import React from "react";
-
-interface RampIconProps {
-  color: string;
-}
-
-const RampIcon: React.FC<RampIconProps> = ({ color }) => {
-  return (
-    <svg
-      width="41"
-      height="40"
-      viewBox="0 0 41 40"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <g>
-        <path
-          d="M5.87534 12.2375C6.01655 12.3457 6.13305 12.4827 6.21706 12.6395C6.30106 12.7963 6.35064 12.9692 6.36247 13.1467C6.3743 13.3242 6.34812 13.5021 6.28567 13.6687C6.22322 13.8353 6.12595 13.9866 6.00034 14.1125C5.88336 14.2288 5.79055 14.3671 5.72726 14.5194C5.66397 14.6717 5.63146 14.8351 5.63159 15C5.63069 15.1569 5.67902 15.3101 5.76978 15.4381C5.86053 15.5661 5.98914 15.6624 6.13749 15.7135C6.28585 15.7645 6.44649 15.7678 6.59679 15.7228C6.74709 15.6777 6.87949 15.5867 6.97534 15.4625C9.67534 12.175 14.2253 10 19.3816 10H19.5441C20.3126 10.0121 21.0716 9.82934 21.7503 9.46876L22.8503 8.87501C24.6855 7.96409 26.7078 7.49332 28.7566 7.50001H30.4128C30.6751 7.49944 30.9326 7.57001 31.158 7.70422C31.3833 7.83843 31.568 8.03125 31.6925 8.26214C31.8169 8.49303 31.8764 8.75334 31.8645 9.01535C31.8527 9.27736 31.77 9.53126 31.6253 9.75001L30.0066 12.175C29.8311 12.4427 29.7585 12.7649 29.8022 13.082C29.8459 13.3991 30.0028 13.6897 30.2441 13.9C31.0527 14.6066 31.7707 15.4106 32.3816 16.2938C32.6276 16.6623 32.9602 16.9649 33.3503 17.1752C33.7403 17.3855 34.176 17.497 34.6191 17.5H35.6691C36.3222 17.5 36.9485 17.7595 37.4103 18.2213C37.8722 18.6831 38.1316 19.3094 38.1316 19.9625V22.0875C38.1282 22.5768 37.9813 23.0544 37.7092 23.461C37.437 23.8676 37.0514 24.1854 36.6003 24.375L36.4003 24.4625C34.9951 25.0489 33.8389 26.1078 33.1316 27.4563C32.0967 29.3929 30.5924 31.039 28.7566 32.2438C28.5572 32.3842 28.3957 32.5719 28.2867 32.79C28.1776 33.0082 28.1243 33.25 28.1316 33.4938V35.625C28.1577 36.0875 28.0026 36.5419 27.6992 36.892C27.3959 37.242 26.9681 37.4601 26.5066 37.5H23.8503C23.5135 37.4847 23.1894 37.3661 22.9222 37.1604C22.655 36.9547 22.4575 36.6718 22.3566 36.35L22.1191 35.725C22.0336 35.4852 21.8717 35.2801 21.6582 35.1412C21.4448 35.0023 21.1918 34.9374 20.9378 34.9563C20.4253 34.9854 19.9066 35 19.3816 35C18.4356 35.0014 17.4911 34.9283 16.5566 34.7813C16.438 34.7634 16.3169 34.7722 16.2021 34.8071C16.0873 34.842 15.9818 34.902 15.8932 34.9829C15.8045 35.0637 15.7351 35.1633 15.6898 35.2744C15.6446 35.3855 15.6247 35.5053 15.6316 35.625C15.6577 36.0875 15.5026 36.5419 15.1992 36.892C14.8959 37.242 14.4681 37.4601 14.0066 37.5H11.3503C11.0135 37.4847 10.6894 37.3661 10.4222 37.1604C10.155 36.9547 9.95754 36.6718 9.85659 36.35L5.72534 27.6625C5.52559 27.2977 5.34615 26.9221 5.18784 26.5375V26.475C4.30762 24.3142 4.1713 21.9219 4.80034 19.675C4.89639 19.3156 4.89552 18.9371 4.79781 18.5781C4.7001 18.2191 4.50905 17.8924 4.24409 17.6313C3.57012 16.9466 3.18465 16.0293 3.16721 15.0687C3.14977 14.1081 3.50168 13.1775 4.15034 12.4688C4.36076 12.2273 4.65186 12.0707 4.96928 12.0282C5.2867 11.9856 5.60876 12.06 5.87534 12.2375ZM28.2566 19.9563C28.5038 19.9563 28.7455 19.8829 28.9511 19.7456C29.1566 19.6082 29.3168 19.413 29.4114 19.1846C29.5061 18.9562 29.5308 18.7049 29.4826 18.4624C29.4343 18.2199 29.3153 17.9972 29.1405 17.8224C28.9657 17.6476 28.7429 17.5285 28.5005 17.4803C28.258 17.432 28.0066 17.4568 27.7782 17.5514C27.5498 17.646 27.3546 17.8062 27.2173 18.0118C27.0799 18.2174 27.0066 18.459 27.0066 18.7063C27.0066 19.0378 27.1383 19.3557 27.3727 19.5901C27.6071 19.8246 27.9251 19.9563 28.2566 19.9563ZM14.6628 14.2188C14.6963 14.3797 14.7612 14.5325 14.8538 14.6684C14.9464 14.8042 15.0649 14.9205 15.2025 15.0104C15.3401 15.1004 15.4941 15.1624 15.6557 15.1928C15.8172 15.2232 15.9832 15.2214 16.1441 15.1875C18.3027 14.6941 20.5411 14.6622 22.7128 15.0938C23.0369 15.165 23.376 15.1047 23.6556 14.9259C23.9351 14.7472 24.1322 14.4647 24.2035 14.1406C24.2747 13.8166 24.2144 13.4775 24.0356 13.1979C23.8569 12.9184 23.5744 12.7213 23.2503 12.65C20.7331 12.1402 18.1364 12.17 15.6316 12.7375C15.3077 12.8065 15.0243 13.0013 14.8439 13.279C14.6634 13.5567 14.6005 13.8947 14.6691 14.2188H14.6628Z"
-          fill={color}
-        />
-        <path
-          d="M19.8253 8.0125C18.3031 8.0125 17.069 6.77848 17.069 5.25625C17.069 3.73402 18.3031 2.5 19.8253 2.5C21.3475 2.5 22.5815 3.73402 22.5815 5.25625C22.5815 6.77848 21.3475 8.0125 19.8253 8.0125Z"
-          fill={color}
-        />
-      </g>
-    </svg>
-  );
-};
-
-export default RampIcon;
