--- components/icons/SendIcon.tsx
@@ -1,28 +1,0 @@
-import React from "react";
-
-interface SendIconProps {
-  color: string;
-}
-
-const SendIcon: React.FC<SendIconProps> = ({ color }) => {
-  return (
-    <svg
-      width="34"
-      height="34"
-      viewBox="0 0 34 34"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M22.6872 24.5C22.6872 23.77 22.3972 23.07 21.8811 22.5538C21.3649 22.0377 20.6649 21.7477 19.9349 21.7477L10.0679 21.7477L11.7812 20.0344C12.037 19.7787 12.2398 19.4751 12.3782 19.141C12.5166 18.8069 12.5878 18.4488 12.5878 18.0871C12.5878 17.7255 12.5166 17.3674 12.3782 17.0333C12.2398 16.6992 12.037 16.3956 11.7812 16.1399C11.5255 15.8842 11.2219 15.6813 10.8878 15.5429C10.5537 15.4046 10.1956 15.3333 9.83398 15.3333C9.47234 15.3333 9.11425 15.4046 8.78014 15.5429C8.44603 15.6813 8.14245 15.8842 7.88673 16.1399L1.47388 22.5527C1.21798 22.8084 1.01497 23.1119 0.876464 23.446C0.737956 23.7802 0.666663 24.1383 0.666663 24.5C0.666663 24.8617 0.737956 25.2198 0.876464 25.554C1.01497 25.8881 1.21798 26.1916 1.47388 26.4472L7.88674 32.8601C8.14245 33.1158 8.44603 33.3186 8.78014 33.457C9.11425 33.5954 9.47235 33.6667 9.83398 33.6667C10.1956 33.6667 10.5537 33.5954 10.8878 33.457C11.2219 33.3186 11.5255 33.1158 11.7812 32.8601C12.037 32.6044 12.2398 32.3008 12.3782 31.9667C12.5166 31.6326 12.5878 31.2745 12.5878 30.9128C12.5878 30.5512 12.5166 30.1931 12.3782 29.859C12.2398 29.5249 12.037 29.2213 11.7812 28.9656L10.0679 27.2523L19.9349 27.2523C20.6649 27.2523 21.3649 26.9623 21.8811 26.4462C22.3972 25.93 22.6872 25.2299 22.6872 24.5Z"
-        fill={color}
-      />
-      <path
-        d="M11.9794 9.49999C11.9794 10.2299 12.2694 10.93 12.7855 11.4462C13.3017 11.9623 14.0018 12.2523 14.7317 12.2523L24.5987 12.2523L22.8854 13.9656C22.6297 14.2213 22.4268 14.5249 22.2885 14.859C22.1501 15.1931 22.0788 15.5512 22.0788 15.9128C22.0788 16.2745 22.1501 16.6326 22.2885 16.9667C22.4268 17.3008 22.6297 17.6044 22.8854 17.8601C23.1411 18.1158 23.4447 18.3186 23.7788 18.457C24.1129 18.5954 24.471 18.6667 24.8327 18.6667C25.1943 18.6667 25.5524 18.5954 25.8865 18.457C26.2206 18.3186 26.5242 18.1158 26.7799 17.8601L33.1928 11.4472C33.4487 11.1916 33.6517 10.8881 33.7902 10.554C33.9287 10.2198 34 9.86168 34 9.49999C34 9.1383 33.9287 8.78015 33.7902 8.44603C33.6517 8.1119 33.4487 7.80836 33.1928 7.55274L26.7799 1.1399C26.5242 0.884186 26.2206 0.68134 25.8865 0.54295C25.5524 0.404557 25.1943 0.333325 24.8327 0.333325C24.471 0.333325 24.1129 0.404557 23.7788 0.542951C23.4447 0.68134 23.1411 0.884183 22.8854 1.1399C22.6297 1.39562 22.4268 1.6992 22.2885 2.03331C22.1501 2.36742 22.0788 2.72551 22.0788 3.08715C22.0788 3.44878 22.1501 3.80688 22.2885 4.14099C22.4268 4.4751 22.6297 4.77868 22.8854 5.03439L24.5987 6.7477L14.7317 6.74769C14.0018 6.74769 13.3017 7.03767 12.7855 7.55383C12.2694 8.06998 11.9794 8.77003 11.9794 9.49999Z"
-        fill={color}
-      />
-    </svg>
-  );
-};
-
-export default SendIcon;
