--- components/icons/WalletIcon.tsx
@@ -1,24 +1,0 @@
-import React from "react";
-
-interface WalletIconProps {
-  color: string;
-}
-
-const WalletIcon: React.FC<WalletIconProps> = ({ color }) => {
-  return (
-    <svg
-      width="36"
-      height="30"
-      viewBox="0 0 36 30"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M31.75 5H6.75C6.56633 5.00268 6.38433 4.96484 6.21695 4.88919C6.04956 4.81353 5.90091 4.70192 5.78157 4.56228C5.66222 4.42265 5.57512 4.25843 5.52645 4.08131C5.47778 3.90418 5.46875 3.71851 5.5 3.5375C5.56315 3.23614 5.73073 2.96678 5.97316 2.77695C6.21559 2.58712 6.51729 2.48903 6.825 2.5H31.75C32.0815 2.5 32.3995 2.3683 32.6339 2.13388C32.8683 1.89946 33 1.58152 33 1.25C33 0.918479 32.8683 0.600537 32.6339 0.366117C32.3995 0.131696 32.0815 0 31.75 0H4.625C3.532 0.00329717 2.4847 0.438953 1.71183 1.21183C0.938953 1.9847 0.503297 3.032 0.5 4.125V24.2937C0.504943 25.8056 1.10772 27.2542 2.17678 28.3232C3.24584 29.3923 4.69438 29.9951 6.20625 30H29.7937C31.3056 29.9951 32.7542 29.3923 33.8232 28.3232C34.8923 27.2542 35.4951 25.8056 35.5 24.2937V8.75C35.5 7.75544 35.1049 6.80161 34.4016 6.09835C33.6984 5.39509 32.7446 5 31.75 5ZM28 20C27.5055 20 27.0222 19.8534 26.6111 19.5787C26.2 19.304 25.8795 18.9135 25.6903 18.4567C25.5011 17.9999 25.4516 17.4972 25.548 17.0123C25.6445 16.5273 25.8826 16.0819 26.2322 15.7322C26.5819 15.3826 27.0273 15.1445 27.5123 15.048C27.9972 14.9516 28.4999 15.0011 28.9567 15.1903C29.4135 15.3795 29.804 15.7 30.0787 16.1111C30.3534 16.5222 30.5 17.0055 30.5 17.5C30.5 18.163 30.2366 18.7989 29.7678 19.2678C29.2989 19.7366 28.663 20 28 20Z"
-        fill={color}
-      />
-    </svg>
-  );
-};
-
-export default WalletIcon;
