+++ components/icons/icon-map.tsx
@@ -1,0 +1,222 @@
+import React from "react";
+
+interface AppsIconProps {
+  color: string;
+}
+
+const AppsIcon: React.FC<AppsIconProps> = ({ color }) => {
+  return (
+    <svg
+      width="35"
+      height="35"
+      viewBox="0 0 35 35"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M20.9 12.5H14.1C13.2163 12.5 12.5 13.2163 12.5 14.1V20.9C12.5 21.7837 13.2163 22.5 14.1 22.5H20.9C21.7837 22.5 22.5 21.7837 22.5 20.9V14.1C22.5 13.2163 21.7837 12.5 20.9 12.5Z"
+        fill={color}
+      />
+      <path
+        d="M20.9 25H14.1C13.2163 25 12.5 25.7163 12.5 26.6V33.4C12.5 34.2837 13.2163 35 14.1 35H20.9C21.7837 35 22.5 34.2837 22.5 33.4V26.6C22.5 25.7163 21.7837 25 20.9 25Z"
+        fill={color}
+      />
+      <path
+        d="M27.925 0H26.6C26.1757 0 25.7687 0.168571 25.4686 0.468629C25.1686 0.768687 25 1.17565 25 1.6V8.4C25 8.82435 25.1686 9.23131 25.4686 9.53137C25.7687 9.83143 26.1757 10 26.6 10H33.4C33.8243 10 34.2313 9.83143 34.5314 9.53137C34.8314 9.23131 35 8.82435 35 8.4V7.075C35 5.19859 34.2546 3.39904 32.9278 2.07222C31.601 0.7454 29.8014 0 27.925 0Z"
+        fill={color}
+      />
+      <path
+        d="M8.4 12.5H1.6C0.716344 12.5 0 13.2163 0 14.1V20.9C0 21.7837 0.716344 22.5 1.6 22.5H8.4C9.28366 22.5 10 21.7837 10 20.9V14.1C10 13.2163 9.28366 12.5 8.4 12.5Z"
+        fill={color}
+      />
+      <path
+        d="M7.075 35H8.40625C8.82951 34.9983 9.23487 34.829 9.53358 34.5292C9.83229 34.2293 10 33.8233 10 33.4V26.6C10 26.1757 9.83143 25.7687 9.53137 25.4686C9.23131 25.1686 8.82435 25 8.4 25H1.6C1.17565 25 0.768687 25.1686 0.468629 25.4686C0.168571 25.7687 0 26.1757 0 26.6V27.9312C0.00165665 29.8066 0.747787 31.6045 2.07443 32.93C3.40107 34.2555 5.19968 35 7.075 35Z"
+        fill={color}
+      />
+      <path
+        d="M1.6 10H8.4C8.82435 10 9.23131 9.83143 9.53137 9.53137C9.83143 9.23131 10 8.82435 10 8.4V1.6C10 1.17565 9.83143 0.768687 9.53137 0.468629C9.23131 0.168571 8.82435 0 8.4 0H7.075C5.19859 0 3.39904 0.7454 2.07222 2.07222C0.7454 3.39904 0 5.19859 0 7.075V8.40625C0.00165338 8.82951 0.170954 9.23487 0.470833 9.53358C0.770712 9.83229 1.17673 10 1.6 10Z"
+        fill={color}
+      />
+      <path
+        d="M33.4 25H26.6C26.1757 25 25.7687 25.1686 25.4686 25.4686C25.1686 25.7687 25 26.1757 25 26.6V33.4C25 33.8243 25.1686 34.2313 25.4686 34.5314C25.7687 34.8314 26.1757 35 26.6 35H27.9312C29.8066 34.9983 31.6045 34.2522 32.93 32.9256C34.2555 31.5989 35 29.8003 35 27.925V26.6C35 26.1757 34.8314 25.7687 34.5314 25.4686C34.2313 25.1686 33.8243 25 33.4 25Z"
+        fill={color}
+      />
+      <path
+        d="M20.9 0H14.1C13.2163 0 12.5 0.716344 12.5 1.6V8.4C12.5 9.28366 13.2163 10 14.1 10H20.9C21.7837 10 22.5 9.28366 22.5 8.4V1.6C22.5 0.716344 21.7837 0 20.9 0Z"
+        fill={color}
+      />
+      <path
+        d="M33.4 12.5H26.6C25.7163 12.5 25 13.2163 25 14.1V20.9C25 21.7837 25.7163 22.5 26.6 22.5H33.4C34.2837 22.5 35 21.7837 35 20.9V14.1C35 13.2163 34.2837 12.5 33.4 12.5Z"
+        fill={color}
+      />
+    </svg>
+  );
+};
+
+const DollarIcon = () => {
+  return (
+    <svg
+      width="24"
+      height="25"
+      viewBox="0 0 24 25"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <rect y="0.5" width="24" height="24" rx="12" fill="white" />
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M24 12.5C24 19.15 18.65 24.5 12 24.5C5.34996 24.5 0 19.15 0 12.5C0 5.84996 5.34996 0.5 12 0.5C18.65 0.5 24 5.84996 24 12.5ZM5.00004 10.0499C3.54996 13.9 5.55 18.25 9.45 19.6499C9.6 19.75 9.75 19.9499 9.75 20.0999V20.8C9.75 20.9 9.75 20.95 9.69996 20.9999C9.65004 21.2 9.45 21.2999 9.24996 21.2C6.45 20.3 4.29996 18.1499 3.39996 15.35C1.89996 10.6 4.5 5.54992 9.24996 4.04992C9.3 4 9.39996 4 9.45 4C9.65004 4.04992 9.75 4.19992 9.75 4.39996V5.09992C9.75 5.35 9.65004 5.5 9.45 5.59996C7.40004 6.34996 5.75004 7.94992 5.00004 10.0499ZM14.3 4.24996C14.35 4.04992 14.55 3.94996 14.75 4.04992C17.5 4.94992 19.7 7.09996 20.6 9.94996C22.1 14.6999 19.5 19.75 14.75 21.25C14.7 21.2999 14.6 21.2999 14.55 21.2999C14.35 21.25 14.25 21.1 14.25 20.9V20.2C14.25 19.9499 14.35 19.7999 14.55 19.7C16.6 18.95 18.25 17.35 19 15.25C20.45 11.3999 18.45 7.04992 14.55 5.65C14.4 5.54992 14.25 5.35 14.25 5.14996V4.45C14.25 4.34992 14.25 4.3 14.3 4.24996ZM12.15 11.8C14.25 12.05 15.3 12.65 15.3 14.3999C15.3 15.7499 14.3 16.7999 12.8 17.05V18.25C12.75 18.5 12.6 18.65 12.4 18.65H11.65C11.4 18.5999 11.25 18.4499 11.25 18.25V17.05C9.6 16.7999 8.79996 15.8999 8.60004 14.65V14.6C8.60004 14.3999 8.75004 14.2499 8.94996 14.2499H9.80004C9.95004 14.2499 10.1 14.35 10.15 14.5499C10.3 15.2999 10.75 15.85 12.05 15.85C13 15.85 13.7 15.3 13.7 14.5C13.7 13.7 13.25 13.4 11.85 13.15C9.75 12.8999 8.75004 12.25 8.75004 10.6C8.75004 9.34996 9.69996 8.35 11.2 8.14996V7C11.25 6.74992 11.4 6.59992 11.6 6.59992H12.35C12.6 6.64996 12.75 6.79996 12.75 7V8.2C13.9 8.29996 14.8 9.14992 15 10.25V10.3C15 10.4999 14.85 10.6499 14.65 10.6499H13.85C13.7 10.6499 13.55 10.55 13.5 10.4C13.25 9.64996 12.75 9.34996 11.85 9.34996C10.85 9.34996 10.35 9.79984 10.35 10.4999C10.35 11.2 10.65 11.6 12.15 11.8Z"
+        fill="#2775CA"
+      />
+    </svg>
+  );
+};
+
+const MaticIcon = () => {
+  return (
+    <svg
+      width="24"
+      height="25"
+      viewBox="0 0 24 25"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <rect y="0.5" width="24" height="24" rx="12" fill="white" />
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M24 12.5C24 19.1274 18.6274 24.5 12 24.5C5.37258 24.5 0 19.1274 0 12.5C0 5.87258 5.37258 0.5 12 0.5C18.6274 0.5 24 5.87258 24 12.5ZM16.0345 19.358L15.9325 19.3994L15.8186 19.5H15.5246C15.1969 19.5 15.1532 19.4767 15.0762 19.4358L15.0746 19.4349C14.9966 19.3935 11.679 17.4527 11.625 17.4113C11.5335 17.3446 11.4526 17.2648 11.385 17.1746C11.3198 17.0864 11.2692 16.9885 11.2351 16.8846C11.2284 16.862 11.2228 16.8431 11.218 16.824C11.1931 16.7233 11.1931 16.6164 11.1931 15.9201V15.1568L11.979 14.7012L12.7649 14.2456V15.003C12.7706 15.791 12.7709 15.7919 12.8122 15.9175C12.8143 15.9239 12.8166 15.9307 12.8189 15.9379C12.8893 16.14 13.0186 16.317 13.1909 16.4468L14.1567 17.0385C14.2874 17.1192 14.4062 17.1931 14.5125 17.2592C14.8892 17.4935 15.1092 17.6302 15.1466 17.6302C15.2834 17.6904 15.4317 17.7215 15.5815 17.7215C15.7314 17.7215 15.8796 17.6904 16.0165 17.6302C16.1545 17.571 17.8942 16.5414 17.9542 16.5C18.1695 16.3407 18.3136 16.1052 18.3562 15.8432V13.4231C18.3142 13.1684 18.1772 12.9384 17.9722 12.7781C17.9182 12.7189 16.1125 11.6302 16.0405 11.5947C16.0028 11.5737 15.9625 11.5578 15.9205 11.5474C15.661 11.4525 15.3741 11.4631 15.1226 11.5769L13.9227 12.3047C12.8129 12.9734 12.6329 13.074 11.979 13.4349C11.8897 13.4842 11.807 13.5297 11.7281 13.5731C11.229 13.8475 10.8856 14.0364 10.0412 14.5473L8.8414 15.2515C8.69779 15.318 8.54109 15.3524 8.38245 15.3524C8.2238 15.3524 8.06712 15.318 7.92351 15.2515C7.83952 15.2101 4.50994 13.2751 4.43194 13.2278C4.32325 13.1462 4.22811 13.0484 4.14998 12.9379C4.07419 12.8274 4.02307 12.7024 4 12.571V10.3994V8.22781C4.02294 8.09964 4.07131 7.97722 4.14234 7.86755C4.21338 7.75788 4.30569 7.66315 4.41395 7.58876C4.44394 7.56509 5.23585 7.10356 6.17773 6.55918C7.44957 5.82545 7.91151 5.56508 7.9775 5.5355L8.06749 5.5H8.66742H8.72141L8.84738 5.55917C9.00336 5.64201 12.267 7.53551 12.327 7.57693C12.497 7.70625 12.626 7.88072 12.6989 8.07987C12.7096 8.1126 12.7181 8.13772 12.725 8.16328C12.7529 8.26713 12.7529 8.37847 12.7529 9.03847V9.77811L11.967 10.2337L11.1691 10.6893V9.94379V9.20415V9.11538C11.1457 9.02403 11.1093 8.93639 11.0611 8.85502C10.9871 8.72227 10.8822 8.60872 10.7551 8.52367C10.6232 8.44674 8.88339 7.43492 8.8234 7.40533C8.68417 7.34213 8.53271 7.30941 8.37944 7.30941C8.22618 7.30941 8.07474 7.34213 7.9355 7.40533C7.90401 7.42949 7.30315 7.78375 6.79351 8.08422C6.42934 8.29893 6.11174 8.48618 6.08174 8.50591C5.98698 8.57078 5.90386 8.65079 5.83577 8.74261C5.7677 8.83947 5.71322 8.94496 5.67379 9.05622C5.59766 9.51189 5.58354 9.97555 5.63179 10.4349C5.61036 10.8291 5.61036 11.2241 5.63179 11.6183C5.66213 11.7532 5.717 11.8815 5.79377 11.9971C5.86292 12.0914 5.94581 12.1751 6.03974 12.2456C6.13573 12.3166 7.92951 13.3758 7.9955 13.3758C8.13377 13.4347 8.28284 13.4648 8.43346 13.4645C8.58553 13.4663 8.73578 13.4319 8.87138 13.3639C8.91338 13.3521 9.47131 13.0444 10.0712 12.6775L12.015 11.5533L13.9827 10.4112C14.2533 10.2551 14.4743 10.1253 14.6507 10.0217C15.0023 9.81534 15.1766 9.71302 15.2126 9.71302C15.3474 9.6527 15.4936 9.6215 15.6415 9.6215C15.7895 9.6215 15.9358 9.6527 16.0705 9.71302L19.2801 11.5651C19.4443 11.6481 19.5958 11.7537 19.73 11.8787C19.8734 12.0294 19.9674 12.219 20 12.4231V16.6539C19.9727 16.8379 19.889 17.0092 19.76 17.145C19.6873 17.2231 19.6022 17.289 19.5081 17.3403L17.7442 18.3639L16.0345 19.358Z"
+        fill="#854CE6"
+      />
+    </svg>
+  );
+};
+
+const ExportIcon = () => {
+  return (
+    <svg
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M23.2494 18.3756L23.2494 14.7643C23.2494 14.4292 23.1162 14.1077 22.8792 13.8707C22.6422 13.6337 22.3208 13.5006 21.9856 13.5006L20.7594 13.5006C20.4249 13.5016 20.1044 13.6352 19.8682 13.8721C19.632 14.1089 19.4994 14.4298 19.4994 14.7643L19.4994 16.7406C19.4984 17.0751 19.3648 17.3956 19.1279 17.6318C18.891 17.868 18.5702 18.0006 18.2356 18.0006L7.26314 18.0006C6.92797 18.0006 6.60653 17.8674 6.36953 17.6304C6.13253 17.3934 5.99939 17.072 5.99939 16.7368L5.99939 14.7643C5.99939 14.4292 5.86625 14.1077 5.62925 13.8707C5.39225 13.6337 5.07081 13.5006 4.73564 13.5006L3.50939 13.5006C3.17487 13.5016 2.85439 13.6352 2.61821 13.8721C2.38202 14.1089 2.24939 14.4298 2.24939 14.7643L2.24939 18.3756C2.24939 19.2707 2.60497 20.1291 3.2379 20.7621C3.87084 21.395 4.72928 21.7506 5.62439 21.7506L19.8744 21.7506C20.7695 21.7506 21.6279 21.395 22.2609 20.7621C22.8938 20.1291 23.2494 19.2707 23.2494 18.3756Z"
+        fill="#80858E"
+      />
+      <path
+        d="M7.38693 6.99059C7.34826 7.02924 7.31759 7.07514 7.29667 7.12565C7.27574 7.17615 7.26497 7.23029 7.26497 7.28496C7.26497 7.33964 7.27574 7.39377 7.29667 7.44428C7.31759 7.49479 7.34826 7.54069 7.38693 7.57934L8.92443 9.11309C8.96308 9.15175 9.00897 9.18243 9.05948 9.20335C9.10999 9.22428 9.16413 9.23505 9.2188 9.23505C9.27348 9.23505 9.32761 9.22428 9.37812 9.20335C9.42863 9.18243 9.47452 9.15175 9.51318 9.11309L10.9794 7.64684C11.002 7.62489 11.0304 7.61 11.0613 7.60402C11.0922 7.59804 11.1242 7.60123 11.1532 7.61319C11.1823 7.62515 11.2073 7.64535 11.2251 7.67132C11.2428 7.69729 11.2526 7.72789 11.2532 7.75934L11.2532 15.3343C11.2532 15.4441 11.2965 15.5494 11.3738 15.6274C11.451 15.7053 11.5559 15.7496 11.6657 15.7506L13.8332 15.7506C13.9436 15.7506 14.0494 15.7067 14.1275 15.6287C14.2056 15.5506 14.2494 15.4447 14.2494 15.3343L14.2494 7.75934C14.25 7.72789 14.2598 7.69729 14.2775 7.67132C14.2953 7.64535 14.3203 7.62515 14.3494 7.61319C14.3785 7.60123 14.4104 7.59804 14.4413 7.60402C14.4722 7.61 14.5006 7.62489 14.5232 7.64684L15.9894 9.11309C16.0281 9.15175 16.074 9.18243 16.1245 9.20335C16.175 9.22428 16.2291 9.23505 16.2838 9.23505C16.3385 9.23505 16.3926 9.22428 16.4431 9.20335C16.4936 9.18243 16.5395 9.15175 16.5782 9.11309L18.1119 7.57559C18.1506 7.53694 18.1813 7.49104 18.2022 7.44053C18.2231 7.39002 18.2339 7.33589 18.2339 7.28122C18.2339 7.22654 18.2231 7.17241 18.2022 7.1219C18.1813 7.07139 18.1506 7.02549 18.1119 6.98684L13.8107 2.68934C13.6714 2.54987 13.5059 2.43924 13.3238 2.36375C13.1417 2.28826 12.9466 2.24941 12.7494 2.24941C12.5523 2.24941 12.3571 2.28826 12.175 2.36375C11.9929 2.43924 11.8275 2.54987 11.6882 2.68934L7.38693 6.99059Z"
+        fill="#80858E"
+      />
+    </svg>
+  );
+};
+
+const ImportIcon = () => {
+  return (
+    <svg
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M6.43875 22.5L17.5612 22.5C18.8708 22.499 20.1264 21.9784 21.0524 21.0524C21.9784 20.1264 22.499 18.8708 22.5 17.5612L22.5 6.43875C22.499 5.12922 21.9784 3.8736 21.0524 2.94762C20.1264 2.02164 18.8708 1.50099 17.5613 1.5C17.1634 1.5 16.7819 1.65803 16.5006 1.93934C16.2193 2.22064 16.0613 2.60217 16.0613 3C16.0613 3.39782 16.2193 3.77936 16.5006 4.06066C16.7819 4.34196 17.1634 4.5 17.5613 4.5C17.8159 4.5 18.068 4.55015 18.3032 4.64758C18.5384 4.74501 18.7521 4.88782 18.9322 5.06785C19.2957 5.43143 19.5 5.92456 19.5 6.43875L19.5 16.5L4.5 16.5L4.5 6.43875C4.5 5.92456 4.70426 5.43143 5.06785 5.06785C5.43143 4.70426 5.92456 4.5 6.43875 4.5C6.83658 4.5 7.21811 4.34196 7.49941 4.06066C7.78072 3.77935 7.93875 3.39782 7.93875 3C7.93875 2.60217 7.78072 2.22064 7.49941 1.93934C7.21811 1.65803 6.83658 1.5 6.43875 1.5C5.12922 1.50099 3.87361 2.02164 2.94763 2.94762C2.02164 3.8736 1.501 5.12921 1.5 6.43875L1.5 17.5612C1.50099 18.8708 2.02164 20.1264 2.94762 21.0524C3.8736 21.9784 5.12922 22.499 6.43875 22.5Z"
+        fill="#80858E"
+      />
+      <path
+        d="M12 1.5C11.6022 1.5 11.2207 1.65804 10.9394 1.93934C10.6581 2.22064 10.5 2.60217 10.5 3L10.5 8.3775L9.56627 7.44375C9.4269 7.30438 9.26145 7.19383 9.07936 7.11841C8.89727 7.04298 8.70211 7.00417 8.50502 7.00417C8.30792 7.00417 8.11276 7.04298 7.93067 7.11841C7.74858 7.19383 7.58313 7.30438 7.44377 7.44375C7.3044 7.58311 7.19385 7.74856 7.11843 7.93065C7.043 8.11274 7.00418 8.30791 7.00418 8.505C7.00418 8.70209 7.043 8.89725 7.11843 9.07934C7.19385 9.26143 7.3044 9.42688 7.44377 9.56625L10.9388 13.0612C11.0781 13.2007 11.2435 13.3114 11.4256 13.3868C11.6077 13.4623 11.8029 13.5012 12 13.5012C12.1971 13.5012 12.3923 13.4623 12.5744 13.3868C12.7565 13.3114 12.922 13.2007 13.0613 13.0612L16.5563 9.56625C16.6956 9.42688 16.8062 9.26143 16.8816 9.07934C16.957 8.89725 16.9959 8.70209 16.9959 8.505C16.9959 8.30791 16.957 8.11275 16.8816 7.93066C16.8062 7.74857 16.6956 7.58311 16.5563 7.44375C16.4169 7.30438 16.2515 7.19383 16.0694 7.11841C15.8873 7.04299 15.6921 7.00417 15.495 7.00417C15.2979 7.00417 15.1028 7.04299 14.9207 7.11841C14.7386 7.19383 14.5731 7.30438 14.4338 7.44375L13.5 8.3775L13.5 3C13.5 2.60217 13.342 2.22064 13.0607 1.93934C12.7794 1.65804 12.3978 1.5 12 1.5Z"
+        fill="#80858E"
+      />
+    </svg>
+  );
+};
+
+interface RampIconProps {
+  color: string;
+}
+
+const RampIcon: React.FC<RampIconProps> = ({ color }) => {
+  return (
+    <svg
+      width="41"
+      height="40"
+      viewBox="0 0 41 40"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <g>
+        <path
+          d="M5.87534 12.2375C6.01655 12.3457 6.13305 12.4827 6.21706 12.6395C6.30106 12.7963 6.35064 12.9692 6.36247 13.1467C6.3743 13.3242 6.34812 13.5021 6.28567 13.6687C6.22322 13.8353 6.12595 13.9866 6.00034 14.1125C5.88336 14.2288 5.79055 14.3671 5.72726 14.5194C5.66397 14.6717 5.63146 14.8351 5.63159 15C5.63069 15.1569 5.67902 15.3101 5.76978 15.4381C5.86053 15.5661 5.98914 15.6624 6.13749 15.7135C6.28585 15.7645 6.44649 15.7678 6.59679 15.7228C6.74709 15.6777 6.87949 15.5867 6.97534 15.4625C9.67534 12.175 14.2253 10 19.3816 10H19.5441C20.3126 10.0121 21.0716 9.82934 21.7503 9.46876L22.8503 8.87501C24.6855 7.96409 26.7078 7.49332 28.7566 7.50001H30.4128C30.6751 7.49944 30.9326 7.57001 31.158 7.70422C31.3833 7.83843 31.568 8.03125 31.6925 8.26214C31.8169 8.49303 31.8764 8.75334 31.8645 9.01535C31.8527 9.27736 31.77 9.53126 31.6253 9.75001L30.0066 12.175C29.8311 12.4427 29.7585 12.7649 29.8022 13.082C29.8459 13.3991 30.0028 13.6897 30.2441 13.9C31.0527 14.6066 31.7707 15.4106 32.3816 16.2938C32.6276 16.6623 32.9602 16.9649 33.3503 17.1752C33.7403 17.3855 34.176 17.497 34.6191 17.5H35.6691C36.3222 17.5 36.9485 17.7595 37.4103 18.2213C37.8722 18.6831 38.1316 19.3094 38.1316 19.9625V22.0875C38.1282 22.5768 37.9813 23.0544 37.7092 23.461C37.437 23.8676 37.0514 24.1854 36.6003 24.375L36.4003 24.4625C34.9951 25.0489 33.8389 26.1078 33.1316 27.4563C32.0967 29.3929 30.5924 31.039 28.7566 32.2438C28.5572 32.3842 28.3957 32.5719 28.2867 32.79C28.1776 33.0082 28.1243 33.25 28.1316 33.4938V35.625C28.1577 36.0875 28.0026 36.5419 27.6992 36.892C27.3959 37.242 26.9681 37.4601 26.5066 37.5H23.8503C23.5135 37.4847 23.1894 37.3661 22.9222 37.1604C22.655 36.9547 22.4575 36.6718 22.3566 36.35L22.1191 35.725C22.0336 35.4852 21.8717 35.2801 21.6582 35.1412C21.4448 35.0023 21.1918 34.9374 20.9378 34.9563C20.4253 34.9854 19.9066 35 19.3816 35C18.4356 35.0014 17.4911 34.9283 16.5566 34.7813C16.438 34.7634 16.3169 34.7722 16.2021 34.8071C16.0873 34.842 15.9818 34.902 15.8932 34.9829C15.8045 35.0637 15.7351 35.1633 15.6898 35.2744C15.6446 35.3855 15.6247 35.5053 15.6316 35.625C15.6577 36.0875 15.5026 36.5419 15.1992 36.892C14.8959 37.242 14.4681 37.4601 14.0066 37.5H11.3503C11.0135 37.4847 10.6894 37.3661 10.4222 37.1604C10.155 36.9547 9.95754 36.6718 9.85659 36.35L5.72534 27.6625C5.52559 27.2977 5.34615 26.9221 5.18784 26.5375V26.475C4.30762 24.3142 4.1713 21.9219 4.80034 19.675C4.89639 19.3156 4.89552 18.9371 4.79781 18.5781C4.7001 18.2191 4.50905 17.8924 4.24409 17.6313C3.57012 16.9466 3.18465 16.0293 3.16721 15.0687C3.14977 14.1081 3.50168 13.1775 4.15034 12.4688C4.36076 12.2273 4.65186 12.0707 4.96928 12.0282C5.2867 11.9856 5.60876 12.06 5.87534 12.2375ZM28.2566 19.9563C28.5038 19.9563 28.7455 19.8829 28.9511 19.7456C29.1566 19.6082 29.3168 19.413 29.4114 19.1846C29.5061 18.9562 29.5308 18.7049 29.4826 18.4624C29.4343 18.2199 29.3153 17.9972 29.1405 17.8224C28.9657 17.6476 28.7429 17.5285 28.5005 17.4803C28.258 17.432 28.0066 17.4568 27.7782 17.5514C27.5498 17.646 27.3546 17.8062 27.2173 18.0118C27.0799 18.2174 27.0066 18.459 27.0066 18.7063C27.0066 19.0378 27.1383 19.3557 27.3727 19.5901C27.6071 19.8246 27.9251 19.9563 28.2566 19.9563ZM14.6628 14.2188C14.6963 14.3797 14.7612 14.5325 14.8538 14.6684C14.9464 14.8042 15.0649 14.9205 15.2025 15.0104C15.3401 15.1004 15.4941 15.1624 15.6557 15.1928C15.8172 15.2232 15.9832 15.2214 16.1441 15.1875C18.3027 14.6941 20.5411 14.6622 22.7128 15.0938C23.0369 15.165 23.376 15.1047 23.6556 14.9259C23.9351 14.7472 24.1322 14.4647 24.2035 14.1406C24.2747 13.8166 24.2144 13.4775 24.0356 13.1979C23.8569 12.9184 23.5744 12.7213 23.2503 12.65C20.7331 12.1402 18.1364 12.17 15.6316 12.7375C15.3077 12.8065 15.0243 13.0013 14.8439 13.279C14.6634 13.5567 14.6005 13.8947 14.6691 14.2188H14.6628Z"
+          fill={color}
+        />
+        <path
+          d="M19.8253 8.0125C18.3031 8.0125 17.069 6.77848 17.069 5.25625C17.069 3.73402 18.3031 2.5 19.8253 2.5C21.3475 2.5 22.5815 3.73402 22.5815 5.25625C22.5815 6.77848 21.3475 8.0125 19.8253 8.0125Z"
+          fill={color}
+        />
+      </g>
+    </svg>
+  );
+};
+
+interface SendIconProps {
+  color: string;
+}
+
+const SendIcon: React.FC<SendIconProps> = ({ color }) => {
+  return (
+    <svg
+      width="34"
+      height="34"
+      viewBox="0 0 34 34"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M22.6872 24.5C22.6872 23.77 22.3972 23.07 21.8811 22.5538C21.3649 22.0377 20.6649 21.7477 19.9349 21.7477L10.0679 21.7477L11.7812 20.0344C12.037 19.7787 12.2398 19.4751 12.3782 19.141C12.5166 18.8069 12.5878 18.4488 12.5878 18.0871C12.5878 17.7255 12.5166 17.3674 12.3782 17.0333C12.2398 16.6992 12.037 16.3956 11.7812 16.1399C11.5255 15.8842 11.2219 15.6813 10.8878 15.5429C10.5537 15.4046 10.1956 15.3333 9.83398 15.3333C9.47234 15.3333 9.11425 15.4046 8.78014 15.5429C8.44603 15.6813 8.14245 15.8842 7.88673 16.1399L1.47388 22.5527C1.21798 22.8084 1.01497 23.1119 0.876464 23.446C0.737956 23.7802 0.666663 24.1383 0.666663 24.5C0.666663 24.8617 0.737956 25.2198 0.876464 25.554C1.01497 25.8881 1.21798 26.1916 1.47388 26.4472L7.88674 32.8601C8.14245 33.1158 8.44603 33.3186 8.78014 33.457C9.11425 33.5954 9.47235 33.6667 9.83398 33.6667C10.1956 33.6667 10.5537 33.5954 10.8878 33.457C11.2219 33.3186 11.5255 33.1158 11.7812 32.8601C12.037 32.6044 12.2398 32.3008 12.3782 31.9667C12.5166 31.6326 12.5878 31.2745 12.5878 30.9128C12.5878 30.5512 12.5166 30.1931 12.3782 29.859C12.2398 29.5249 12.037 29.2213 11.7812 28.9656L10.0679 27.2523L19.9349 27.2523C20.6649 27.2523 21.3649 26.9623 21.8811 26.4462C22.3972 25.93 22.6872 25.2299 22.6872 24.5Z"
+        fill={color}
+      />
+      <path
+        d="M11.9794 9.49999C11.9794 10.2299 12.2694 10.93 12.7855 11.4462C13.3017 11.9623 14.0018 12.2523 14.7317 12.2523L24.5987 12.2523L22.8854 13.9656C22.6297 14.2213 22.4268 14.5249 22.2885 14.859C22.1501 15.1931 22.0788 15.5512 22.0788 15.9128C22.0788 16.2745 22.1501 16.6326 22.2885 16.9667C22.4268 17.3008 22.6297 17.6044 22.8854 17.8601C23.1411 18.1158 23.4447 18.3186 23.7788 18.457C24.1129 18.5954 24.471 18.6667 24.8327 18.6667C25.1943 18.6667 25.5524 18.5954 25.8865 18.457C26.2206 18.3186 26.5242 18.1158 26.7799 17.8601L33.1928 11.4472C33.4487 11.1916 33.6517 10.8881 33.7902 10.554C33.9287 10.2198 34 9.86168 34 9.49999C34 9.1383 33.9287 8.78015 33.7902 8.44603C33.6517 8.1119 33.4487 7.80836 33.1928 7.55274L26.7799 1.1399C26.5242 0.884186 26.2206 0.68134 25.8865 0.54295C25.5524 0.404557 25.1943 0.333325 24.8327 0.333325C24.471 0.333325 24.1129 0.404557 23.7788 0.542951C23.4447 0.68134 23.1411 0.884183 22.8854 1.1399C22.6297 1.39562 22.4268 1.6992 22.2885 2.03331C22.1501 2.36742 22.0788 2.72551 22.0788 3.08715C22.0788 3.44878 22.1501 3.80688 22.2885 4.14099C22.4268 4.4751 22.6297 4.77868 22.8854 5.03439L24.5987 6.7477L14.7317 6.74769C14.0018 6.74769 13.3017 7.03767 12.7855 7.55383C12.2694 8.06998 11.9794 8.77003 11.9794 9.49999Z"
+        fill={color}
+      />
+    </svg>
+  );
+};
+
+interface WalletIconProps {
+  color: string;
+}
+
+const WalletIcon: React.FC<WalletIconProps> = ({ color }) => {
+  return (
+    <svg
+      width="36"
+      height="30"
+      viewBox="0 0 36 30"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M31.75 5H6.75C6.56633 5.00268 6.38433 4.96484 6.21695 4.88919C6.04956 4.81353 5.90091 4.70192 5.78157 4.56228C5.66222 4.42265 5.57512 4.25843 5.52645 4.08131C5.47778 3.90418 5.46875 3.71851 5.5 3.5375C5.56315 3.23614 5.73073 2.96678 5.97316 2.77695C6.21559 2.58712 6.51729 2.48903 6.825 2.5H31.75C32.0815 2.5 32.3995 2.3683 32.6339 2.13388C32.8683 1.89946 33 1.58152 33 1.25C33 0.918479 32.8683 0.600537 32.6339 0.366117C32.3995 0.131696 32.0815 0 31.75 0H4.625C3.532 0.00329717 2.4847 0.438953 1.71183 1.21183C0.938953 1.9847 0.503297 3.032 0.5 4.125V24.2937C0.504943 25.8056 1.10772 27.2542 2.17678 28.3232C3.24584 29.3923 4.69438 29.9951 6.20625 30H29.7937C31.3056 29.9951 32.7542 29.3923 33.8232 28.3232C34.8923 27.2542 35.4951 25.8056 35.5 24.2937V8.75C35.5 7.75544 35.1049 6.80161 34.4016 6.09835C33.6984 5.39509 32.7446 5 31.75 5ZM28 20C27.5055 20 27.0222 19.8534 26.6111 19.5787C26.2 19.304 25.8795 18.9135 25.6903 18.4567C25.5011 17.9999 25.4516 17.4972 25.548 17.0123C25.6445 16.5273 25.8826 16.0819 26.2322 15.7322C26.5819 15.3826 27.0273 15.1445 27.5123 15.048C27.9972 14.9516 28.4999 15.0011 28.9567 15.1903C29.4135 15.3795 29.804 15.7 30.0787 16.1111C30.3534 16.5222 30.5 17.0055 30.5 17.5C30.5 18.163 30.2366 18.7989 29.7678 19.2678C29.2989 19.7366 28.663 20 28 20Z"
+        fill={color}
+      />
+    </svg>
+  );
+};
+
+const iconMap = {
+  apps: AppsIcon,
+  dollar: DollarIcon,
+  matic: MaticIcon,
+  export: ExportIcon,
+  import: ImportIcon,
+  ramp: RampIcon,
+  send: SendIcon,
+  wallet: WalletIcon
+};
+
+export default iconMap;
